+++ src/cli/constants.evaluate.js
@@ -1,12 +1,12 @@
 import { outdent } from "outdent";
-import { coreOptions } from "./prettier-internal.js";
+import { optionCategories } from "./prettier-internal.js";
 
 const categoryOrder = [
-  coreOptions.CATEGORY_OUTPUT,
-  coreOptions.CATEGORY_FORMAT,
-  coreOptions.CATEGORY_CONFIG,
-  coreOptions.CATEGORY_EDITOR,
-  coreOptions.CATEGORY_OTHER,
+  optionCategories.CATEGORY_OUTPUT,
+  optionCategories.CATEGORY_FORMAT,
+  optionCategories.CATEGORY_CONFIG,
+  optionCategories.CATEGORY_EDITOR,
+  optionCategories.CATEGORY_OTHER,
 ];
 
 const usageSummary = outdent`
