+++ src/language-graphql/options.js
@@ -1,4 +1,4 @@
-import commonOptions from "../common/common-options.js";
+import commonOptions from "../common/common-options.evaluate.js";
 
 // format based on https://github.com/prettier/prettier/blob/main/src/main/core-options.evaluate.js
 const options = {
