+++ src/main/core-options.evaluate.js
@@ -1,12 +1,5 @@
 import { outdent } from "outdent";
-
-const CATEGORY_CONFIG = "Config";
-const CATEGORY_EDITOR = "Editor";
-const CATEGORY_FORMAT = "Format";
-const CATEGORY_OTHER = "Other";
-const CATEGORY_OUTPUT = "Output";
-const CATEGORY_GLOBAL = "Global";
-const CATEGORY_SPECIAL = "Special";
+import optionCategories from "./option-categories.js";
 
 /**
  * @typedef {Object} OptionInfo
@@ -46,7 +39,7 @@ const CATEGORY_SPECIAL = "Special";
 /** @type {{ [name: string]: OptionInfo }} */
 const options = {
   cursorOffset: {
-    category: CATEGORY_SPECIAL,
+    category: optionCategories.CATEGORY_SPECIAL,
     type: "int",
     default: -1,
     range: { start: -1, end: Number.POSITIVE_INFINITY, step: 1 },
@@ -54,10 +47,10 @@ const options = {
       Print (to stderr) where a cursor at the given position would move to after formatting.
       This option cannot be used with --range-start and --range-end.
     `,
-    cliCategory: CATEGORY_EDITOR,
+    cliCategory: optionCategories.CATEGORY_EDITOR,
   },
   endOfLine: {
-    category: CATEGORY_GLOBAL,
+    category: optionCategories.CATEGORY_GLOBAL,
     type: "choice",
     default: "lf",
     description: "Which end of line characters to apply.",
@@ -86,23 +79,23 @@ const options = {
     ],
   },
   filepath: {
-    category: CATEGORY_SPECIAL,
+    category: optionCategories.CATEGORY_SPECIAL,
     type: "path",
     description:
       "Specify the input filepath. This will be used to do parser inference.",
     cliName: "stdin-filepath",
-    cliCategory: CATEGORY_OTHER,
+    cliCategory: optionCategories.CATEGORY_OTHER,
     cliDescription: "Path to the file to pretend that stdin comes from.",
   },
   insertPragma: {
-    category: CATEGORY_SPECIAL,
+    category: optionCategories.CATEGORY_SPECIAL,
     type: "boolean",
     default: false,
     description: "Insert @format pragma into file's first docblock comment.",
-    cliCategory: CATEGORY_OTHER,
+    cliCategory: optionCategories.CATEGORY_OTHER,
   },
   parser: {
-    category: CATEGORY_GLOBAL,
+    category: optionCategories.CATEGORY_GLOBAL,
     type: "choice",
     default: undefined,
     description: "Which parser to use.",
@@ -138,19 +131,19 @@ const options = {
     type: "path",
     array: true,
     default: [{ value: [] }],
-    category: CATEGORY_GLOBAL,
+    category: optionCategories.CATEGORY_GLOBAL,
     description:
       "Add a plugin. Multiple plugins can be passed as separate `--plugin`s.",
     exception: (value) =>
       typeof value === "string" || typeof value === "object",
     cliName: "plugin",
-    cliCategory: CATEGORY_CONFIG,
+    cliCategory: optionCategories.CATEGORY_CONFIG,
   },
   pluginSearchDirs: {
     type: "path",
     array: true,
     default: [{ value: [] }],
-    category: CATEGORY_GLOBAL,
+    category: optionCategories.CATEGORY_GLOBAL,
     description: outdent`
       Custom directory that contains prettier plugins in node_modules subdirectory.
       Overrides default behavior when plugins are searched relatively to the location of Prettier.
@@ -159,17 +152,17 @@ const options = {
     exception: (value) =>
       typeof value === "string" || typeof value === "object",
     cliName: "plugin-search-dir",
-    cliCategory: CATEGORY_CONFIG,
+    cliCategory: optionCategories.CATEGORY_CONFIG,
   },
   printWidth: {
-    category: CATEGORY_GLOBAL,
+    category: optionCategories.CATEGORY_GLOBAL,
     type: "int",
     default: 80,
     description: "The line length where Prettier will try wrap.",
     range: { start: 0, end: Number.POSITIVE_INFINITY, step: 1 },
   },
   rangeEnd: {
-    category: CATEGORY_SPECIAL,
+    category: optionCategories.CATEGORY_SPECIAL,
     type: "int",
     default: Number.POSITIVE_INFINITY,
     range: { start: 0, end: Number.POSITIVE_INFINITY, step: 1 },
@@ -178,10 +171,10 @@ const options = {
       The range will extend forwards to the end of the selected statement.
       This option cannot be used with --cursor-offset.
     `,
-    cliCategory: CATEGORY_EDITOR,
+    cliCategory: optionCategories.CATEGORY_EDITOR,
   },
   rangeStart: {
-    category: CATEGORY_SPECIAL,
+    category: optionCategories.CATEGORY_SPECIAL,
     type: "int",
     default: 0,
     range: { start: 0, end: Number.POSITIVE_INFINITY, step: 1 },
@@ -190,33 +183,33 @@ const options = {
       The range will extend backwards to the start of the first line containing the selected statement.
       This option cannot be used with --cursor-offset.
     `,
-    cliCategory: CATEGORY_EDITOR,
+    cliCategory: optionCategories.CATEGORY_EDITOR,
   },
   requirePragma: {
-    category: CATEGORY_SPECIAL,
+    category: optionCategories.CATEGORY_SPECIAL,
     type: "boolean",
     default: false,
     description: outdent`
       Require either '@prettier' or '@format' to be present in the file's first docblock comment
       in order for it to be formatted.
     `,
-    cliCategory: CATEGORY_OTHER,
+    cliCategory: optionCategories.CATEGORY_OTHER,
   },
   tabWidth: {
     type: "int",
-    category: CATEGORY_GLOBAL,
+    category: optionCategories.CATEGORY_GLOBAL,
     default: 2,
     description: "Number of spaces per indentation level.",
     range: { start: 0, end: Number.POSITIVE_INFINITY, step: 1 },
   },
   useTabs: {
-    category: CATEGORY_GLOBAL,
+    category: optionCategories.CATEGORY_GLOBAL,
     type: "boolean",
     default: false,
     description: "Indent with tabs instead of spaces.",
   },
   embeddedLanguageFormatting: {
-    category: CATEGORY_GLOBAL,
+    category: optionCategories.CATEGORY_GLOBAL,
     type: "choice",
     default: "auto",
     description:
@@ -235,13 +228,4 @@ const options = {
   },
 };
 
-export {
-  CATEGORY_CONFIG,
-  CATEGORY_EDITOR,
-  CATEGORY_FORMAT,
-  CATEGORY_OTHER,
-  CATEGORY_OUTPUT,
-  CATEGORY_GLOBAL,
-  CATEGORY_SPECIAL,
-  options,
-};
+export default options;
