+++ src/main/option-categories.js
@@ -1,0 +1,17 @@
+const CATEGORY_CONFIG = "Config";
+const CATEGORY_EDITOR = "Editor";
+const CATEGORY_FORMAT = "Format";
+const CATEGORY_OTHER = "Other";
+const CATEGORY_OUTPUT = "Output";
+const CATEGORY_GLOBAL = "Global";
+const CATEGORY_SPECIAL = "Special";
+
+export default {
+  CATEGORY_CONFIG,
+  CATEGORY_EDITOR,
+  CATEGORY_FORMAT,
+  CATEGORY_OTHER,
+  CATEGORY_OUTPUT,
+  CATEGORY_GLOBAL,
+  CATEGORY_SPECIAL,
+};
